Abstract out common methods in *CommentEntity
