+++ app/src/Talk/TalkCommentEntity.php
@@ -1,36 +1,14 @@
 <?php
 namespace Talk;
 
-class TalkCommentEntity
+use Application\BaseCommentEntity;
+
+class TalkCommentEntity extends BaseCommentEntity
 {
-    private $data;
-
-    /**
-     * Create new TalkCommentEntity
-     *
-     * @param Object $data Model data retrieved from API
-     */
-    public function __construct($data)
+    public function __construct(\stdClass $data)
     {
-        $this->data = $data;
-    }
-
-    public function getRating()
-    {
-        if (!isset($this->data->rating)) {
-            return null;
-        }
-
-        return $this->data->rating;
-    }
-
-    public function getUserDisplayName()
-    {
-        if (!isset($this->data->user_display_name)) {
-            return null;
-        }
-
-        return $this->data->user_display_name;
+        parent::__construct($data);
+        $this->commentUri = isset($this->data->uri) ? $this->data->uri : null;
     }
 
     public function getUsername()
@@ -40,42 +18,6 @@ class TalkCommentEntity
         }
 
         return $this->data->username;
-    }
-
-    public function getGravatarHash()
-    {
-        if (!isset($this->data->gravatar_hash)) {
-            return null;
-        }
-
-        return $this->data->gravatar_hash;
-    }
-
-    public function getCommentDate()
-    {
-        if (!isset($this->data->created_date)) {
-            return null;
-        }
-
-        return $this->data->created_date;
-    }
-
-    public function getComment()
-    {
-        if (!isset($this->data->comment)) {
-            return null;
-        }
-
-        return $this->data->comment;
-    }
-
-    public function getCommentSource()
-    {
-        if (!isset($this->data->source)) {
-            return null;
-        }
-
-        return $this->data->source;
     }
 
     public function getTalkTitle()
@@ -94,34 +36,6 @@ class TalkCommentEntity
         }
 
         return $this->data->talk_uri;
-    }
-
-    public function getCommentHash()
-    {
-        if (!isset($this->data->uri)) {
-            return null;
-        }
-
-        $hash = md5($this->data->uri);
-        return (substr($hash, 0, 6));
-    }
-
-    public function getReportedUri()
-    {
-        if (!isset($this->data->reported_uri)) {
-            return null;
-        }
-
-        return $this->data->reported_uri;
-    }
-
-    public function getUserUri()
-    {
-        if (!isset($this->data->user_uri)) {
-            return null;
-        }
-
-        return $this->data->user_uri;
     }
 
     public function canRateTalk($user_uri)
